Enforce int-is-boundary-only across the overlay kernels
Convert the 131 internal bare-int violations flagged by
test/check_int_boundary.py to the overlay convention: numeric
sizes/indices/strides -> ptrdiff_t, logical flags -> bool. Fixed-width
Fortran-ABI int stays only in common/ (facade + blas_omp).

- Mechanical/non-hot (flags->bool, small counters, CPU-capability
  flags): kind10 etpmv/etri_kernel/etrsm_kernel/ytpmv/ytpsv, kind16
  qgemm/qsymv/xhemv/qsyrk/qsyr2k/qtrmv/xtrmv, multifloats mf_cpu/
  mf_dispatch. Flags feed branches; blkrange is cold per-call setup;
  unroll counters fully unroll -> codegen-neutral.

- kind16 xtrsm L-side packed port (xl3_complex.{c,h}, xtri_driver.c,
  xtrsm_parallel.c): the OpenBLAS ZTRSM port landed with int throughout.
  Driver blocking/loop-control -> ptrdiff_t, flags -> bool; drop the now-
  needless (int) call casts and the INT_MAX truncation guard. Perf-
  verified via an interleaved dual-link A/B over the packed-path cells
  (L-side Trans/ConjTrans, M>=128): par stays at parity with ob at both
  OMP=1 and OMP=4 (worst new/ob 1.008 / 1.003), no regression vs the int
  baseline. Fuzz passes on xtrsm/xtrmv across all five seeds.

- CI: add a `ctest -L lint` step to ci.yml (before the fuzz rotation)
  and release.yml (test slot) so the boundary guard, which was labeled
  "lint" and never run by the fuzz-only gate, now blocks regressions.

+++ src/epblas-parallel/kind10/ytpsv.c
@@ -224,7 +224,7 @@ static void ytpsv_axpy_col(TC *restrict y_, const TC *restrict a_, TC t, ptrdiff
     const long double *restrict a = (const long double *)a_;
     ptrdiff_t i = 0;
     for (; i + 3 < m; i += 4) {
-        for (int u = 0; u < 4; ++u) {
+        for (ptrdiff_t u = 0; u < 4; ++u) {
             const long double ar = a[2 * (i + u)], ai = a[2 * (i + u) + 1];
             y[2 * (i + u)]     -= tr * ar - ti * ai;
             y[2 * (i + u) + 1] -= tr * ai + ti * ar;
