+++ src/epblas-parallel/kind16/qsyr2k_serial.c
@@ -198,7 +198,7 @@ void qsyr2k_trans_col(ptrdiff_t j, char UPLO, ptrdiff_t n, ptrdiff_t k,
     const TR *Aj = a + j * lda;
     const TR *Bj = b + j * ldb;
     TR *cj = c + j * ldc;
-    const int bmode = (beta == 0.0Q) ? 0 : (beta == 1.0Q) ? 1 : 2;
+    const ptrdiff_t bmode = (beta == 0.0Q) ? 0 : (beta == 1.0Q) ? 1 : 2;
     for (ptrdiff_t i = i_lo; i < i_hi; ++i) {
         const TR *Ai = a + i * lda;
         const TR *Bi = b + i * ldb;
